fix:#196 내정보 확인 > label 좀 더 명확한 이름으로 수정
