+++ src/features/my-page/my-info-content.tsx
@@ -1,18 +1,30 @@
 import clsx from 'clsx';
-import type { FieldList } from '@/features/my-page/my-info';
 import Button from '@/components/ui/button';
 import { useModalStore } from '@/store/use-modal-store';
 import { MODAL_ID } from '@/constants/modal-id-constants';
+import { USER_META_DATA_KEY } from '@/constants/user-meta-data-constants';
 import type { UserMetaDataType } from '@/types/user-meta-data-type';
 
 type Props = {
-  fieldList: FieldList[];
   data: UserMetaDataType;
+};
+type FieldList = {
+  key: keyof UserMetaDataType;
+  label: string;
 };
 
 const { USER_META_DATA } = MODAL_ID;
+const { EXPERIENCE_NAME, REQUIRED_EDUCATION_NAME, JOB_MID_CODE_NAME, LOCATION_NAME, ETC } = USER_META_DATA_KEY;
 
-const MyInfoContent = ({ fieldList, data }: Props) => {
+const fieldList: FieldList[] = [
+  { key: EXPERIENCE_NAME, label: '관련 경력' },
+  { key: REQUIRED_EDUCATION_NAME, label: '최종 학력' },
+  { key: JOB_MID_CODE_NAME, label: '지원 직무' },
+  { key: LOCATION_NAME, label: '근무 지역' },
+  { key: ETC, label: '기타 커리어' },
+];
+
+const MyInfoContent = ({ data }: Props) => {
   const toggleModal = useModalStore((state) => state.toggleModal);
   return (
     <div>
